refactor: 오픈마켓 네트워크 구현방식 수정
