+++ OpenMarket/OpenMarket/Model/Data/ImageCacheManager.swift
@@ -2,7 +2,7 @@
 //  ImageCacheManager.swift
 //  OpenMarket
 //
-//  Created by 김동용 on 2022/07/23.
+//  Created by groot, bard on 2022/07/23.
 //
 
 import UIKit
