+++ OpenMarket/OpenMarket/Model/JSONModel/ProductDetail.swift
@@ -50,9 +50,9 @@ extension ProductDetail {
     }
     
     func pushThumbnailImageCache() {
-        let request = OpenMarketRequest(method: .get, baseURL: self.thumbnail)
+        var request = OpenMarketRequest(method: .get, baseURL: self.thumbnail)
         let session = MyURLSession()
-        session.execute(with: request) { (result: Result<Data, Error>) in
+        session.execute(with: request.SetGetImageRequest()) { (result: Result<Data, Error>) in
             switch result {
             case .success(let success):
                 guard let image = UIImage(data: success) else { return }
